Revert "moved non UI parts to org.palladiosimulator.analyzer.workflow.core"
This reverts commit 661562088fb7febb9bdaf99e7fd1b6775b2676a7.
