+++ bundles/org.palladiosimulator.analyzer.accuracy/src/org/palladiosimulator/analyzer/accuracy/jobs/ShowAccuracyInfluenceAnalysisErrorsJob.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import org.apache.log4j.Logger;
 import org.eclipse.core.runtime.IProgressMonitor;
-import org.palladiosimulator.analyzer.workflow.core.jobs.IIssueReceiver;
+import org.palladiosimulator.analyzer.workflow.jobs.IIssueReceiver;
 
 import de.uka.ipd.sdq.errorhandling.core.SeverityAndIssue;
 import de.uka.ipd.sdq.errorhandling.dialogs.issues.DisplayIssuesDialog;
