+++ bundles/org.palladiosimulator.analyzer.quality.editor/src-man/org/palladiosimulator/analyzer/quality/presentation/SelectRepositoryWizardPage.java
@@ -12,7 +12,7 @@ import org.eclipse.swt.events.ModifyListener;
 import org.eclipse.swt.layout.GridLayout;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Text;
-import org.palladiosimulator.analyzer.workflow.core.ConstantsContainer;
+import org.palladiosimulator.analyzer.workflow.ConstantsContainer;
 
 import de.uka.ipd.sdq.workflow.launchconfig.tabs.TabHelper;
