+++ bundles/org.palladiosimulator.analyzer.workflow/META-INF/MANIFEST.MF
@@ -12,7 +12,6 @@ Require-Bundle: org.palladiosimulator.pcm;bundle-version="4.0.0",
  de.uka.ipd.sdq.workflow.logging.console;bundle-version="5.2.0";visibility:=reexport,
  de.uka.ipd.sdq.workflow.mdsd;bundle-version="2.0.1",
  de.uka.ipd.sdq.workflow.ui;bundle-version="5.2.0";visibility:=reexport,
- org.palladiosimulator.analyzer.workflow.core;bundle-version="5.2.0";visibility:=reexport,
  org.palladiosimulator.analyzer.completions;bundle-version="1.0.0",
  org.eclipse.ui;visibility:=reexport,
  org.eclipse.core.runtime;visibility:=reexport,
@@ -20,7 +19,8 @@ Require-Bundle: org.palladiosimulator.pcm;bundle-version="4.0.0",
  org.eclipse.debug.core;visibility:=reexport,
  de.uka.ipd.sdq.errorhandling;bundle-version="1.0.0",
  org.eclipse.debug.ui;bundle-version="3.5.1";visibility:=reexport
-Export-Package: org.palladiosimulator.analyzer.workflow.blackboard,
+Export-Package: org.palladiosimulator.analyzer.workflow,
+ org.palladiosimulator.analyzer.workflow.blackboard,
  org.palladiosimulator.analyzer.workflow.configurations,
  org.palladiosimulator.analyzer.workflow.jobs,
  org.palladiosimulator.analyzer.workflow.runconfig
