+++ bundles/org.palladiosimulator.analyzer.workflow/src/org/palladiosimulator/analyzer/workflow/jobs/IIssueReceiver.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.analyzer.workflow.core.jobs;
+package org.palladiosimulator.analyzer.workflow.jobs;
 
 import java.util.List;
